Moved core fdr functions from the script base to the package
These are central functions for setting up properly pruned
random indices and the central fdr calculation engine. At this
point, some minor adjustments and revised documentation of the
original code, but the calculation engine is not yet ready
to go (but we can do indices).

This has alos had major impact on the already existing R6 data
structure in the package, which has been modified to accept
the new code more gracefully.